Consolidate unit tests for MustParseMAC and GetLocalBroadcastIP into ip_test.go
Addressed reviewer feedback to move new unit tests from ip_extra_test.go
into the existing ip_test.go file for better consistency and maintainability.

Changes:
- Moved TestMustParseMAC tests to ip_test.go
- Moved TestGetLocalBroadcastIP tests to ip_test.go
- Removed ip_extra_test.go file

All tests pass successfully with 86.3% coverage on pkg/util/ip package.
