+++ coverage_new.out
@@ -1,0 +1,75 @@
+mode: set
+antrea.io/antrea/pkg/util/ip/ip.go:40.52,42.2 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:47.90,51.41 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:51.41,53.36 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:53.36,55.61 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:55.61,64.19 6 1
+antrea.io/antrea/pkg/util/ip/ip.go:65.10,65.46 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:65.46,70.19 4 0
+antrea.io/antrea/pkg/util/ip/ip.go:74.2,74.22 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:79.66,87.31 6 1
+antrea.io/antrea/pkg/util/ip/ip.go:87.31,89.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:89.8,91.3 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:96.2,97.51 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:97.51,101.31 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:101.31,103.4 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:105.3,106.52 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:108.2,108.23 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:111.53,118.2 5 1
+antrea.io/antrea/pkg/util/ip/ip.go:122.55,124.45 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:124.45,126.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:130.2,130.39 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:130.39,131.44 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:131.44,132.48 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:132.48,139.5 4 1
+antrea.io/antrea/pkg/util/ip/ip.go:142.2,142.19 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:149.55,152.21 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:152.21,154.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:155.2,157.45 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:171.78,172.21 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:173.20,174.16 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:175.20,176.16 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:177.19,178.15 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:179.19,180.15 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:181.22,182.21 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:183.20,184.16 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:185.10,186.22 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:191.44,193.16 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:193.16,194.56 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:196.2,196.14 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:199.48,201.16 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:201.16,202.55 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:204.2,204.13 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:208.49,209.36 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:209.36,211.3 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:212.2,212.36 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:212.36,214.3 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:215.2,215.44 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:215.44,217.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:218.2,218.33 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:218.33,220.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:221.2,221.13 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:232.52,233.36 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:233.36,235.3 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:236.2,238.20 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:238.20,240.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:241.2,241.19 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:241.19,243.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:244.2,244.33 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:244.33,246.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:247.2,247.13 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:250.32,252.25 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:252.25,253.51 1 0
+antrea.io/antrea/pkg/util/ip/ip.go:255.2,255.11 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:261.51,265.2 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:268.52,269.54 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:269.54,271.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:273.2,275.15 2 1
+antrea.io/antrea/pkg/util/ip/ip.go:275.15,277.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:279.2,279.37 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:284.75,288.25 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:288.25,290.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:290.8,292.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:295.2,299.34 3 1
+antrea.io/antrea/pkg/util/ip/ip.go:299.34,301.3 1 1
+antrea.io/antrea/pkg/util/ip/ip.go:303.2,304.19 2 1
